+++ src/getCurVersion.py
@@ -5,7 +5,7 @@ time = str(datetime.today().strftime('%d_%m_%Y-%H_%M_%S'))
 filname = "logs/krillYouBotLog-" + time + ".log"
 
 def getCurVersion():
-    return '2.2 Hotfix 3'
+    return '2.3'
 
 def get_filname():
     return filname